pyproject.toml
[tool.poetry]
name = "tidy3d-webapi"
version = "0.1.0"
description = ""
authors = ["An Li <kidylee@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.2"
toml = "^0.10.2"
pydantic = "^1.10.2"
requests = "^2.28.1"
boto3 = "^1.23.0"
tidy3d-beta = "^1.7.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
autohooks = "^22.8.1"
autohooks-plugin-black = "^22.8.1"
autohooks-plugin-pylint = "^22.8.1"
autohooks-plugin-isort = "^22.8.0"
pylint = "^2.15.0"
pytest-ordering = "^0.6"
Sphinx = "5.2.3"
docutils = "0.17.1"
sphinxcontrib-openapi = "0.7.0"
mistune = "0.8.4"
sphinx-rtd-theme = "^1.0.0"
responses = "^0.22.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"



[tool.isort]
profile = "black"

[tool.black]
line-length = 100
target-version = ["py37", "py38", "py39", "py310", "py311"]
exclude = '''
/(
    \.git
)/
'''

[tool.autohooks]
mode = "poetry"
pre-commit = ["autohooks.plugins.black", "autohooks.plugins.isort", "autohooks.plugins.pylint"]

[tool.autohooks.plugins.pylint]
arguments = ["--rcfile=.pylintrc"]


[tool.poetry.scripts]
tidy3d = "tidy3d_webapi.cli:tidy3d_cli"